requirements.txt
# Google Sheets API
gspread==6.0.0
google-auth==2.27.0

# Supabase
supabase==2.9.0

# Data manipulation
pandas==2.2.0

# Environment variables
python-dotenv==1.0.1

# Utilities
python-dateutil==2.8.2
websockets>=13.0
